restructure library for easier maintenance and usage
