+++ tsconfig.json
@@ -20,16 +20,11 @@
       "jasmine",
       "node"
     ],
-    "lib": ["es2015", "dom"]
+    "lib": ["es2015", "dom"],
+    "outDir": "dist"
   },
   "files": [
-    "index.ts",
-    "src/providers/angulartics2-ga.ts",
-    "src/providers/angulartics2-kissmetrics.ts",
-    "src/providers/angulartics2-mixpanel.ts",
-    "src/providers/angulartics2-piwik.ts",
-    "src/providers/angulartics2-segment.ts",
-    "src/providers/angulartics2-baidu-analytics.ts"
+    "src/index.ts"
   ],
   "compileOnSave": false,
   "angularCompilerOptions": {
